organize model package
moved some stuff into 2 new sub-packages:

- `meta`: annotations, superclasses, and wrappers
- `parts`: components of domain model classes
