+++ build.gradle
@@ -75,7 +75,7 @@ dependencies {
     implementation 'org.slf4j:slf4j-simple:2.0.0-alpha1'
     implementation 'org.apache.commons:commons-lang3:3.11'
 
-    implementation 'com.github.jkatzwinkel:tla-common:etl-model-SNAPSHOT'
+    implementation 'com.github.jkatzwinkel:tla-common:master-SNAPSHOT'
     implementation 'com.github.rosmord.jsesh:jsesh:release-7.5.5'
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test:2.4.0'
