+++ src/main/java/tla/web/model/ThsEntry.java
@@ -8,6 +8,8 @@ import lombok.EqualsAndHashCode;
 import lombok.Singular;
 import tla.domain.model.Language;
 import tla.domain.model.meta.BTSeClass;
+import tla.web.model.meta.BackendPath;
+import tla.web.model.meta.TLAObject;
 
 @Data
 @BackendPath("ths")
