+++ src/main/java/tla/web/model/meta/BackendPath.java
@@ -1,4 +1,4 @@
-package tla.web.model;
+package tla.web.model.meta;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
