+++ src/main/java/tla/web/model/parts/Transcription.java
@@ -1,4 +1,4 @@
-package tla.web.model;
+package tla.web.model.parts;
 
 import lombok.Data;
 import lombok.AllArgsConstructor;
