+++ src/main/java/tla/web/repo/TlaClient.java
@@ -13,9 +13,9 @@ import tla.domain.dto.extern.SearchResultsWrapper;
 import tla.domain.dto.extern.SingleDocumentWrapper;
 import tla.domain.dto.meta.AbstractDto;
 import tla.domain.dto.meta.DocumentDto;
-import tla.web.model.BackendPath;
 import tla.web.model.Lemma;
-import tla.web.model.TLAObject;
+import tla.web.model.meta.BackendPath;
+import tla.web.model.meta.TLAObject;
 import tla.web.model.ThsEntry;
 
 /**
